using wakeupcall instead of wakeupcall_from_file
